Added connection pool, DAO interfaces, DAO work in progress

+++ src/main/java/com/solvd/bankjdbc/dao/IBaseDao.java
@@ -2,7 +2,7 @@ package com.solvd.bankjdbc.dao;
 
 public interface IBaseDao <T>{
     void create (T object);
-    T getById (long id);
+    T getById (int id);
     void update (T object);
-    void remove (long id);
+    void remove (int id);
 }
